requirements.txt
# Core
fastapi==0.110.*
uvicorn[standard]==0.29.*
sqlalchemy==2.0.*
alembic==1.13.*
psycopg2-binary==2.9.*
pydantic[email]==2.6.*
pydantic-settings==2.2.*

# Background jobs
arq==0.25.*
redis==5.0.*

# File storage
boto3==1.34.*

# Email
resend==0.8.*

# Auth
pyjwt==2.8.*
cryptography==42.*

# ElevenLabs
elevenlabs==1.*

# Video processing
ffmpeg-python==0.2.*

# HTTP client
httpx==0.27.*

# Dev
pytest==8.*
pytest-asyncio==0.23.*
